requirements.txt
aiolimiter==1.0.0
anyio==3.6.2
APScheduler==3.10.1
arrow==1.2.3
astroid==2.15.5
attrs==23.1.0
backports.zoneinfo==0.2.1
beautiful-date==2.2.0
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dill==0.3.6
docopt==0.6.2
gcsa==2.1.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.59.0
h11==0.14.0
httpcore==0.17.1
httplib2==0.22.0
httpx==0.24.1
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
oauthlib==3.2.2
pipreqs==0.4.13
platformdirs==3.5.1
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pylint==2.17.4
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==20.3
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
TatSu==5.8.3
tomli==2.0.1
tomlkit==0.11.8
tornado==6.3.2
typing-extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.22.0
wrapt==1.15.0
yarg==0.1.9
